Move the example App.tsx back into place
